Fix energy weapon logic with power cell ammo
