+++ Content.Shared/_N14/Weapons/Ranged/Components/N14EnergyWeaponComponent.cs
@@ -1,6 +1,7 @@
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype;
+using Content.Shared.Weapons.Ranged;
 using Content.Shared.Weapons.Ranged.Components;
 
 namespace Content.Shared._N14.Weapons.Ranged.Components;
